Выполнена задача по написанию механизма уведомления пользователя о его подписке а так же проведен рефакт кода

+++ src/main/java/com/skillbox/cryptobot/dto/SubscriberDto.java
@@ -15,6 +15,6 @@ public class SubscriberDto {
     @NotNull
     private Long botID;
 
-    private Long priceSubscription;
+    private Double priceSubscription;
 
 }
